feat (tests): Move the registerElement and related functions outside of the NS components registry, in order to ease mocking
