+++ platform/nativescript/compiler/modules/router.js
@@ -1,4 +1,4 @@
-import { normalizeElementName } from '../../element-registry'
+import { normalizeElementName } from '../../register'
 import { addAttr } from 'compiler/helpers'
 
 function preTransformNode(el) {
