+++ platform/nativescript/runtime/index.js
@@ -3,8 +3,7 @@ import { warn } from 'core/util/index'
 import { patch } from './patch'
 import { mountComponent } from 'core/instance/lifecycle'
 import { compileToFunctions } from '../compiler/index'
-import { registerElement, getElementMap } from '../element-registry'
-import { makeMap } from 'shared/util'
+import { registerElement, getElementMap } from '../register'
 
 import Vue from 'core/index'
 import DocumentNode from '../renderer/DocumentNode'
